feat(widgets): move widgets to components
