+++ packages/ui/src/components/bar-chart/barChart.vue
