+++ packages/ui/src/components/fitted-actions/fittedActions.vue
@@ -1,12 +1,12 @@
 <script setup lang="ts">
   import type { Ref } from 'vue';
   import type { AcvFittedAction, AcvFittedActionsProps, AcvFittedItem } from './fittedActions.ts';
+  import AcvButton from '@/components/button/button.vue';
+  import AcvDropdown from '@/components/dropdown/dropdown.vue';
+  import AcvDropdownMenu from '@/components/dropdown/dropdownMenu.vue';
+  import useFittedActions from '@/components/fitted-actions/useFittedActions.ts';
+  import ListItem from '@/components/list-item/listItem.vue';
   import { ref, toRefs } from 'vue';
-  import AcvButton from '../../components/button/button.vue';
-  import AcvDropdown from '../../components/dropdown/dropdown.vue';
-  import AcvDropdownMenu from '../../components/dropdown/dropdownMenu.vue';
-  import ListItem from '../../components/list-item/listItem.vue';
-  import useFittedActions from './useFittedActions.ts';
 
   import './fittedActions.css';
