+++ packages/ui/src/components/index.ts
@@ -26,10 +26,12 @@ export * from './disclosure/public.ts';
 export * from './divider/public.ts';
 export * from './drawer/public.ts';
 export * from './dropdown/public.ts';
+export * from './empty/public.ts';
 export * from './file-browser/public.ts';
 export * from './file-picker/public.ts';
 export * from './file-upload/public.ts';
 export * from './filter/public.ts';
+export * from './fitted-actions/public.ts';
 export * from './flex/public.ts';
 export * from './footer/public.ts';
 export * from './form/public.ts';
@@ -44,9 +46,11 @@ export * from './infinite-scroll/public.ts';
 export * from './input/public.ts';
 export * from './input-base/public.ts';
 export * from './input-number/public.ts';
+export * from './line-chart/public.ts';
 export * from './link/public.ts';
 export * from './list/public.ts';
 export * from './list-item/public.ts';
+export * from './loading/public.ts';
 export * from './main/public.ts';
 export * from './map/public.ts';
 export * from './markdown-editor/public.ts';
@@ -60,6 +64,7 @@ export * from './option/public.ts';
 export * from './option-group/public.ts';
 export * from './pagination/public.ts';
 export * from './password/public.ts';
+export * from './pie-chart/public.ts';
 export * from './popover/public.ts';
 export * from './popper/public.ts';
 export * from './progress/public.ts';
@@ -74,6 +79,7 @@ export * from './skeleton/public.ts';
 export * from './slider/public.ts';
 export * from './spinner/public.ts';
 export * from './split-button/public.ts';
+export * from './stacked-area-chart/public.ts';
 export * from './stepper/public.ts';
 export * from './stepper-item/public.ts';
 export * from './submenu/public.ts';
@@ -84,6 +90,7 @@ export * from './tabs/public.ts';
 export * from './tag/public.ts';
 export * from './textarea/public.ts';
 export * from './theme-provider/public.ts';
+export * from './theme-switcher/public.ts';
 export * from './tiles/public.ts';
 export * from './time-picker/public.ts';
 export * from './toolbar/public.ts';
@@ -91,3 +98,4 @@ export * from './tooltip/public.ts';
 export * from './tree/public.ts';
 export * from './upload/public.ts';
 export * from './virtual-scroll/public.ts';
+export * from './widget-wrapper/public.ts';
