+++ packages/ui/src/components/pie-chart/widget.chart.js
@@ -1,9 +1,9 @@
-import { t } from '../../locale';
-import { cssValue } from '../../utils/color-converters';
-import { chartColors, isValidColor } from '../../utils/color-utils';
-import { getLocale } from '../../utils/local-storage';
-import prettySize from '../../utils/prettySize.ts';
-import { getFadedColors } from '../../utils/widget.ts';
+import { t } from '@/locale/index.js';
+import { cssValue } from '@/utils/color-converters.js';
+import { chartColors, isValidColor } from '@/utils/color-utils.ts';
+import { getLocale } from '@/utils/local-storage.js';
+import prettySize from '@/utils/prettySize.ts';
+import { getFadedColors } from '@/utils/widget.ts';
 // import { addResizeListener, removeResizeListener } from '../../utils/resize-event';
 
 const IS_LOADING = 'LOADING';
