+++ packages/ui/src/index.ts
@@ -6,12 +6,10 @@ import type {
 import * as components from './components/index.ts';
 import * as directives from './directives/index.ts';
 import { isObject } from './utils/util.ts';
-import * as widgets from './widgets/index.ts';
 
 export * from './components/index.ts';
 export * from './composables/index.ts';
 export * from './directives/index.ts';
-export * from './widgets/index.ts';
 
 export default {
   install(app: App, _options: never): void {
@@ -22,12 +20,6 @@ export default {
     for (const [name, component] of Object.entries(components as unknown as Record<string, Component>)) {
       if (isObject(component) && 'name' in component) {
         app.component(component.name ?? name, component);
-      }
-    }
-
-    for (const [name, widget] of Object.entries(widgets as unknown as Record<string, Component>)) {
-      if (isObject(widget) && 'name' in widget) {
-        app.component(widget.name ?? name, widget);
       }
     }
   }
