--- packages/ui/src/widgets/index.ts
@@ -1,8 +1,0 @@
-export * from './empty/public.ts';
-export * from './fitted-actions/public.ts';
-export * from './line-chart/public.ts';
-export * from './loading/public.ts';
-export * from './pie-chart/public.ts';
-export * from './stacked-area-chart/public.ts';
-export * from './theme-switcher/public.ts';
-export * from './widget-wrapper/public.ts';
